Caching system/ repositories added
Implemented the weather repository for retrieving the various weather objects from the server/cache.

Created a generic hive repository class that is responsible for getting and saving any object T in cache against a unique key.

Hive requires type adapters to cache objects directly so they have also been generated.

Currently as it stands if there is a Weather/Forecast in cache it will use the cached version. This is not completely correct as we need to check connectivity status to decide whether to fetch from the local storage/ server.