Fixed date bug
There was an issue whereby the sunset/sunrise time was displaying the time relative to GMT. So the sunset could be at 06:00am which is true relative to GMT but not true in relation to that specific timezone. I have now fixed this issue using the timezone value provided by the api.