Themed Text implemented.
Themed text is just a wrapper for the Text widget. It accepts a themed text type enum and styles the text widget accordingly. The theme style is based on the predefined themes we created for light/dark mode (appThemes).

While implementing this change I decided to tidy a few things up such as the folder structure and the location of enums, constants and extensions. Not sure if this is the best structure but it works well for the current weather app so will continue to use this for now.