onresume implementation
AppLifeCycleState resume used to fetch weather after periods of inactivity from user.

If the app is backgrounded and then is resumed say after an hour inactivity the UI does not update as no fetch event is called. However, now if the user opens the weather app from background and the time between the last fetched weather and current time is greater then 20 minutes we refetch the weather.

The API recommends only making calls every 10 minutes, but for the resume method to work efficiently 10 minutes seemed too little. This is because if the time difference exceeds 10 minutes the app fetches the new weather, however i have noticed that after the 10 minute period it is common for the api to not have any new weather data so we are again fetching for no reason. It also meant the weather time is not updated and the app will try to fetch the weather every time the user opens the app until the new data is fetched from the server. 20 minutes is more then enough time to retrieve new data which consistently works and lets be honest a weather app is not going to be opened every 10 minutes, its likely to be open once or twice a day,