Added in error handling and optimised some of my async methods
The Dio error is now passed back in the weather failed state stream. We take this error and convert it to a ui element appropriate for the existing use case. I created a themed button widget; this is a custom elevated button with the correct theming and sizing.

The animation controller is initialised, played and reset based on the weather success state, however it also needs to used for the widget returned from weather error fetch state. I created a void method for completing this functionality and removed repeated code.