Fixed first location caching issue
Replaced bloc builder with bloc consumer which fires search add event on weather fetch success state. Removed stream subscription as it is no longer necessary due to listener. The bug is that streamsubscription was not firing on first launch due to the location of the blocbuilder of type SearchBloc. So either I needed to add blocbuilder of type searchbloc to the widget tree, or add a bloclistener and remove the subscription. the latter seemed more concise.