Modified endpoint to get daily forecast
Daily forecast is not permitted for free API users, so updated to new API that provides this data. However the response does not contain location name and requires geo coords (lat/long) therefore we need to make a seperate API call to get this data. In any case we would be making three calls but now we only make two which is better anyways.

Had to update models and screens to reflect weather model changes.