Added dismissible functionality for listview
Added Dismissible widget and added new parameter in search removed event. The reason for this new parameter is that because our search removed method in our bloc initially yields a search loading state, the whole listview displays a spinner momentarily, so the list appears to flash. If we pass this isDismissed boolean as true we can just bypass this SearchLoading() state and only stream either SearchRemovedSuccessState or a failure state.

- Removed unused imports.